Update from Daniel of the notify plugin. Play with this, it is very cool, it looks at UAVObjects and based on their values it has rules to play sounds. It is flexible and a much better approach than text to speech which was originally considered but still sounds bad and hard to do cross platform.
Sound collection path if you are building in a build directory would be:

../../share/openpilotgcs/sounds

Select default as the language. There is one issue remaining which Daniel is working on, 0.0x values do not play, these are rare ut will be fixed shortly. It is now mature enough to test with however.

Additionally as part of this commit, typeo correct of a sound file.

git-svn-id: svn://svn.openpilot.org/OpenPilot@849 ebee16cc-31ac-478f-84a7-5cbb03baadba