OP-121 OP-122 Flight: Improved USB HID communication.   1. Added reenumeration function and call it on USB init (device will appear after reprogramming now)   2. Moved buffer.c to general flight/Libraries location   3. Removed the 62 byte transmission limitation by adding a transmission buffer   4. Sped up USB communication by increasing endpoint polling frequency
Note, that the nonblocking and blocking USB send functions are not blocking entirely correcting.  The blocking calls the nonblocking, and the nonblocking blocks until the last chunk has started tranmission if it's a big transmission.  The buffering I added would generalize to non-blocking nicely, but would require using the EP1(IN) callback to handle most of the tranmission.  This creates a lot of issues if one function is pushing data onto the buffer and the interrupt is sending.

git-svn-id: svn://svn.openpilot.org/OpenPilot@1403 ebee16cc-31ac-478f-84a7-5cbb03baadba