Use QSettings key/properties instead of byte arrays for plugin configurations.
I've left in the reading of the byte arrays for now, so people can load their old config files, on the next save they'll be converted. This should be removed at some point in the not too far future though, since it's a lot of duplicate code in each plugin.

I've converted all the plugins, it's certainly possible I made a typo somewhere, I tried to test as much as I could, sorry if I broke something though :)


git-svn-id: svn://svn.openpilot.org/OpenPilot@1599 ebee16cc-31ac-478f-84a7-5cbb03baadba