GCS: Make [UAVGadgetManager] also human readable. Makes the ini file a bit more verbose, but debuggable/hackable :) (Not to mention more robust).
Again, I left in support for loading the binary version, so this shouldn't mess up people's configurations. Keep a backup handy anyways :)

In the future, that needs to be removed though.

Anyone with a 'old style' configuration may notice the binary versions are left in the ini file, I can't easily delete them without deleting the new groups also, but they won't cause any problems, and aren't read. Feel free to delete them for a clean ini file :)

I left in the 'DefaultSettings' stuff, but I think it can be removed too ( or whatever it was meant to support should be added :) ).


git-svn-id: svn://svn.openpilot.org/OpenPilot@1664 ebee16cc-31ac-478f-84a7-5cbb03baadba