OP-294 Add logging to scope plugin.
First attempt to add logging to the scope gadget.
Logs a snapshot of data from the scope configuration at the same rate as the update timer set in the options page.
This has configurability via the options page for logging path.
File name is based on the configuration name and the date/time the logging starts.
This uses synchronous file writes in the update timer.

Tested to work on Win XP and Linux (Fedora 14)

git-svn-id: svn://svn.openpilot.org/OpenPilot@2634 ebee16cc-31ac-478f-84a7-5cbb03baadba